requirements.txt
python-binance
pandas
numpy
talib
asyncio
pytest
pytest-asyncio
google-generativeai>=0.3.0
python-dotenv
aiohttp